fix: striped background on tables
The striped background on tables was not visible because the background color of the container matched the striping.

By setting the background of the container to white, the striping became visible.

In addition, the grey background needed to be added to a parent container further up to ensure the grey background extended as the table grew.